Add files via upload
This project is the second task of my UI/UX Internship with Coding Samurai, where I designed a Minimal To-Do List Mobile App. The goal was to create a clean, intuitive, and user-friendly interface for task management. The app includes screens for adding new tasks, viewing the task list, and deleting completed items. The design follows a minimalistic approach with smooth navigation and a focus on productivity. Tools like Figma were used for prototyping, along with Canva and Google Fonts for supporting visuals and typography. This project demonstrates my skills in mobile UI design, interaction flow, and usability-focused layout structuring.